refactor: :recycle: refatora código duplicado para melhor leitura
